[feature] point_can_use_limit 수정

+++ packages/data/lib/dto/login_features.dto.dart
@@ -12,7 +12,7 @@ class LoginFeaturesDto {
   String? name;
 
   @JsonKey(name: 'point_can_use_limit')
-  int? pointCanUseLimit;
+  double? pointCanUseLimit;
 
   LoginFeaturesDto({
     this.key,
